+++ lib/src/model/data/error/remote_config_no_internet_error.dart
@@ -1,0 +1,1 @@
+class RemoteConfigNoInternetError extends Error {}