+++ lib/src/model/data/version_check/version_check_result.dart
@@ -1,0 +1,11 @@
+class VersionCheckResult {
+  final String? updateUrl;
+  final bool isForceUpdate;
+  final String versionNumber;
+
+  VersionCheckResult({
+    required this.updateUrl,
+    required this.isForceUpdate,
+    required this.versionNumber,
+  });
+}