+++ lib/src/model/data/version_check/version_requirements.dart
@@ -1,0 +1,21 @@
+class VersionRequirements {
+  final VersionRequirement latest;
+  final VersionRequirement? min;
+
+  VersionRequirements({
+    required this.latest,
+    this.min,
+  });
+}
+
+class VersionRequirement {
+  final int buildNumber;
+  final String? updateUrl;
+  final String? versionNumber;
+
+  VersionRequirement({
+    required this.buildNumber,
+    this.updateUrl,
+    this.versionNumber,
+  });
+}