+++ lib/src/service/version_check/version_check_service.dart
@@ -1,0 +1,10 @@
+import 'package:impaktfull_architecture/src/model/data/version_check/version_check_result.dart';
+import 'package:impaktfull_architecture/src/model/data/version_check/version_requirements.dart';
+
+abstract class VersionCheckService {
+  Future<VersionRequirements?> getVersionRequirements();
+
+  Future<void> install(VersionCheckResult result);
+
+  void openStore();
+}